fix: correct contributor allocations and more efficient storage

+++ src/SIR.sol
@@ -63,7 +63,7 @@ contract SIR is Staker, SystemControlAccess {
 
             // Calculate the contributor's issuance
             uint256 issuance = (allocation * (SystemConstants.ISSUANCE - SystemConstants.LP_ISSUANCE_FIRST_3_YEARS)) /
-                type(uint56).max;
+                type(uint24).max;
 
             // Update unclaimed rewards
             uint256 timestampNow = block.timestamp >= timestampIssuanceEnd ? timestampIssuanceEnd : block.timestamp;
